+++ lib/home_screen.dart
@@ -366,7 +366,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                 Align(
                                   alignment: Alignment.centerLeft,
                                   child: Text(
-                                    "BETA 04.27",
+                                    "BETA 04.28",
                                     style: GoogleFonts.manrope(
                                       fontSize: Responsive.font(context, 11),
                                       color: Colors.white.withAlpha(80),